requirements.txt
markdown==3.6
pyspellchecker==0.7.2
tkinterweb==3.23.4
